Big change!
I have improved the compiling system so that it now uses a make file,
with instructions inside the Readme file on how to compile, as apposed
to the terminal compiling I was doing.

I have also edited the Cube class so that it now has methods which draw
certain parameters. I have also implemented this into the main.cpp file.
There are still a few more changes which I want to do, such as applying
a texture, but I will get around to that eventually.